BREAKING CHANGE: apply minimize by default when updating document
Fix #13782
