+++ lib/document.js
@@ -33,6 +33,7 @@ const isExclusive = require('./helpers/projection/isExclusive');
 const inspect = require('util').inspect;
 const internalToObjectOptions = require('./options').internalToObjectOptions;
 const markArraySubdocsPopulated = require('./helpers/populate/markArraySubdocsPopulated');
+const minimize = require('./helpers/minimize');
 const mpath = require('mpath');
 const queryhelpers = require('./queryhelpers');
 const utils = require('./utils');
@@ -3937,42 +3938,6 @@ Document.prototype.$toObject = function(options, json) {
 Document.prototype.toObject = function(options) {
   return this.$toObject(options);
 };
-
-/**
- * Minimizes an object, removing undefined values and empty objects
- *
- * @param {Object} object to minimize
- * @return {Object}
- * @api private
- */
-
-function minimize(obj) {
-  const keys = Object.keys(obj);
-  let i = keys.length;
-  let hasKeys;
-  let key;
-  let val;
-
-  while (i--) {
-    key = keys[i];
-    val = obj[key];
-
-    if (utils.isPOJO(val)) {
-      obj[key] = minimize(val);
-    }
-
-    if (undefined === obj[key]) {
-      delete obj[key];
-      continue;
-    }
-
-    hasKeys = true;
-  }
-
-  return hasKeys
-    ? obj
-    : undefined;
-}
 
 /*!
  * Applies virtuals properties to `json`.
