+++ lib/model.js
@@ -65,6 +65,7 @@ const STATES = require('./connectionstate');
 const util = require('util');
 const utils = require('./utils');
 const MongooseBulkWriteError = require('./error/bulkWriteError');
+const minimize = require('./helpers/minimize');
 
 const VERSION_WHERE = 1;
 const VERSION_INC = 2;
@@ -341,7 +342,19 @@ Model.prototype.$__handleSave = function(options, callback) {
     }
 
     _applyCustomWhere(this, where);
-    this[modelCollectionSymbol].updateOne(where, delta[1], saveOptions).then(
+
+    const update = delta[1];
+    if (this.$__schema.options.minimize) {
+      minimize(update);
+      // minimize might leave us with an empty object, which would
+      // lead to MongoDB throwing a "Update document requires atomic operators" error
+      if (Object.keys(update).length === 0) {
+        handleEmptyUpdate.call(this);
+        return;
+      }
+    }
+
+    this[modelCollectionSymbol].updateOne(where, update, saveOptions).then(
       ret => {
         ret.$where = where;
         callback(null, ret);
@@ -353,6 +366,17 @@ Model.prototype.$__handleSave = function(options, callback) {
       }
     );
   } else {
+    handleEmptyUpdate.call(this);
+    return;
+  }
+
+  // store the modified paths before the document is reset
+  this.$__.modifiedPaths = this.modifiedPaths();
+  this.$__reset();
+
+  _setIsNew(this, false);
+
+  function handleEmptyUpdate() {
     const optionsWithCustomValues = Object.assign({}, options, saveOptions);
     const where = this.$__where();
     const optimisticConcurrency = this.$__schema.options.optimisticConcurrency;
@@ -369,14 +393,7 @@ Model.prototype.$__handleSave = function(options, callback) {
         callback(null, { $where: where, matchedCount });
       })
       .catch(callback);
-    return;
   }
-
-  // store the modified paths before the document is reset
-  this.$__.modifiedPaths = this.modifiedPaths();
-  this.$__reset();
-
-  _setIsNew(this, false);
 };
 
 /*!
